+++ CHANGELOG.md
@@ -9,6 +9,8 @@ This changelog will be used to generate documentation on [release notes page](ht
 - Dependency type was updated to reflect the latest developments in Applicaiton Insights Applicaiton Map feature. You can set a new field - `Target`. `CommandName` was renamed to `Data` for consistancy with the Application Analytics schema. `DependencyKind` will never be send any more and will not be set to "Other" by default. Also there are two more constructors for `DependencyTelemetry` item.
 - Type `SessionStateTelemetry` was marked obsolete. Use `IsFirst` flag in `SessionContext` to indicate that the session is just started.
 - Type `PerformanceCounterTelemetry` was marked obsolete. Use `MetricTelemetry` instead.
+- Marked `RequestTelemetry.HttpMethod` as obsolete. Put http verb as part of the name for the better grouping by name and use custom properties to report http verb as a dimension.
+- Marked `RequestTelemetry.StartTime` as obsolete. Use `TimeStamp` instead.
 
 ## Version 2.2.0-beta1